+++ src/main/java/com/example/hits/presentation/request/taskanswer/TaskRateRequestModel.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class TaskRateRequestModel {
 
-    private int rate;
+    private Float rate;
 }
